refactor: split into library and repl for making embedded language
