+++ CMakeLists.txt
@@ -7,6 +7,7 @@ set(CMAKE_C_STANDARD_REQUIRED ON)
 set(CMAKE_C_STANDARD 17)
 
 add_library(snuk_configs INTERFACE)
+target_include_directories(snuk_configs INTERFACE ${PROJECT_SOURCE_DIR}/include)
 
 if(CMAKE_BUILD_TYPE STREQUAL "Debug")
     target_compile_definitions(snuk_configs INTERFACE SNUK_DEBUG)
@@ -24,13 +25,12 @@ target_compile_definitions(snuk_configs INTERFACE
     SNUK_VERSION_PATCH=${PROJECT_VERSION_PATCH}
 )
 
-add_executable(snuk)
-target_link_libraries(snuk PRIVATE snuk_configs)
-
 add_subdirectory(external)
 
 add_subdirectory(docs)
 
 add_subdirectory(src)
+
+add_subdirectory(repl)
 
 add_subdirectory(tests)
