+++ external/CMakeLists.txt
@@ -2,4 +2,4 @@ add_subdirectory(SnLogger)
 add_subdirectory(SnMemory)
 add_subdirectory(SnFile)
 
-target_link_libraries(snuk PRIVATE snlogger snmemory snfile)
+target_link_libraries(snuk_configs INTERFACE snlogger snmemory snfile)
