+++ include/snuk/darray.h
