+++ include/snuk/interpreter/builtins/builtin_bool.h
@@ -1,6 +1,6 @@
 #pragma once
 
 #include "builtins_common.h"
-#include "defines.h"
+#include "snuk/defines.h"
 
 SnukValue builtin_bool_get_member(SnukValue value, SnukStringView field);
