+++ include/snuk/interpreter/builtins/builtin_str.h
@@ -1,6 +1,6 @@
 #pragma once
 
 #include "builtins_common.h"
-#include "defines.h"
+#include "snuk/defines.h"
 
 SnukValue builtin_str_get_member(SnukValue value, SnukStringView field);
