+++ include/snuk/parser/snuk_expr.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "darray.h"
-#include "defines.h"
 #include "parser_common.h"
-#include "string_view.h"
+#include "snuk/darray.h"
+#include "snuk/defines.h"
+#include "snuk/string_view.h"
 
 /**
  * @brief Parser expression node kinds.
