+++ include/snuk/snuk_string.h
