+++ src/interpreter/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(HEADERS
+set(PUBLIC_HEADERS
     interpreter.h
     snuk_signal.h
     snuk_value.h
@@ -6,11 +6,18 @@ set(HEADERS
     snuk_env.h
 )
 
+set(HEADERS
+)
+
 set(SRCS
     interpreter.c
     snuk_value.c
 )
 
-target_sources(snuk PRIVATE ${HEADERS} ${SRCS})
+set(INCLUDE_BASE "${PROJECT_SOURCE_DIR}/include/snuk/interpreter")
+
+list(TRANSFORM PUBLIC_HEADERS PREPEND "${INCLUDE_BASE}/")
+
+target_sources(snuk PRIVATE ${PUBLIC_HEADERS} ${HEADERS} ${SRCS})
 
 add_subdirectory(builtins)
