+++ src/parser/CMakeLists.txt
@@ -1,10 +1,13 @@
-set(HEADERS
-    parser_common.h
+set(PUBLIC_HEADERS
     snuk_item.h
     snuk_expr.h
     snuk_type.h
     parser.h
     snuk_var.h
+    parser_common.h
+)
+
+set(HEADERS
 )
 
 set(SRCS
@@ -15,4 +18,7 @@ set(SRCS
     snuk_var.c
 )
 
-target_sources(snuk PRIVATE ${HEADERS} ${SRCS})
+set(INCLUDE_BASE "${PROJECT_SOURCE_DIR}/include/snuk/parser")
+list(TRANSFORM PUBLIC_HEADERS PREPEND "${INCLUDE_BASE}/")
+
+target_sources(snuk PRIVATE ${PUBLIC_HEADERS} ${HEADERS} ${SRCS})
