+++ tests/unit/CMakeLists.txt
@@ -1,11 +1,11 @@
 function(add_snuk_test name source)
     add_executable(${name}
         ${source}
-        ${CMAKE_SOURCE_DIR}/src/logger.c
-        ${CMAKE_SOURCE_DIR}/src/memory.c
+        ${PROJECT_SOURCE_DIR}/src/logger.c
+        ${PROJECT_SOURCE_DIR}/src/memory.c
     )
     add_dependencies(run_tests ${name})
-    target_include_directories(${name} PRIVATE ${CMAKE_SOURCE_DIR}/src)
+    target_include_directories(${name} PRIVATE ${PROJECT_SOURCE_DIR}/include)
     target_link_libraries(${name} PRIVATE snlogger snmemory snfile)
     add_test(NAME ${name} COMMAND ${name})
     set_tests_properties(${name} PROPERTIES LABELS "unit")
