+++ tests/unit/snuk_string.c
@@ -1,6 +1,6 @@
 #include "test_framework.h"
 
-#include <snuk_string.h>
+#include <snuk/snuk_string.h>
 
 ADD_TEST(test_is_alpha) {
     // lowercase and uppercase letters
